requirements.txt
Django==1.8.4
djangorestframework==3.1.3
django-filter==0.10.0
drf-extensions==0.3.1
uWSGI==2.0.15
pytest==3.1.3
mysqlclient==1.3.7
pandas==0.19.2
django-debug-toolbar==1.7
requests==2.9.1
django-json-field==0.5.7
